+++ src/main/java/com/immobilier/gestionImmobiliere/donnees/localisation/repository/SecteurRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SecteurRepository extends JpaRepository<Secteur, Long> {
-    Page<Secteur> findByVille_IdVille(Long idVille, Pageable pageable);
-    boolean existsByCodeSecteurIgnoreCaseAndVille_IdVille(String codeSecteur, Long idVille);
+public interface SecteurRepository extends JpaRepository<Secteur, Integer> {
+    Page<Secteur> findByVille_IdVille(Integer idVille, Pageable pageable);
+    boolean existsByCodeSecteurIgnoreCaseAndVille_IdVille(String codeSecteur, Integer idVille);
 }
