+++ src/main/java/com/immobilier/gestionImmobiliere/donnees/localisation/repository/VilleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VilleRepository extends JpaRepository<Ville, Long> {
-    Page<Ville> findByPays_IdPays(Long idPays, Pageable pageable);
-    boolean existsByCodeVilleIgnoreCaseAndPays_IdPays(String codeVille, Long idPays);
+public interface VilleRepository extends JpaRepository<Ville, Integer> {
+    Page<Ville> findByPays_IdPays(Integer idPays, Pageable pageable);
+    boolean existsByCodeVilleIgnoreCaseAndPays_IdPays(String codeVille, Integer idPays);
 }
