+++ src/main/java/com/immobilier/gestionImmobiliere/exceptions/PaysNotFoundException.java
@@ -1,7 +1,7 @@
 package com.immobilier.gestionImmobiliere.exceptions;
 
 public class PaysNotFoundException extends RuntimeException {
-    public PaysNotFoundException(Long id) {
+    public PaysNotFoundException(Integer id) {
         super("Aucun localisation trouvé avec l'id : " + id);
     }
 }
